Updated LogIn
LocalStorage and Validations

The localStorage saves the login data to simulate a real login. Everything related to it is extracted to a seperate file. The handling of the side effects happens through asyncThunks and an API file.

The validations are extracted to a sparate file for scalability.

Optimized the SignUp flow.

Tested the app with the 'profiler' to confirm it works in an optimized way. All API requests are done only if needed. All components are rendered only if needed.